change all month datatype to double

+++ src/Profitable.java
@@ -2,7 +2,7 @@
 public interface Profitable 
 {
 	double getInterest();
-	double calculateProfit(double capital , int month);
-	double calculateTotal(double capital, int month);
+	double calculateProfit(double capital , double month);
+	double calculateTotal(double capital, double month);
 	String printCondition();
 }
